move condstore tests to separate file
